amazon-products-app/requirements.txt
flask==3.1.1
databricks-sdk==0.44.0
